// src/actions/types.ts
export enum actionTypes {
    authenticate,
    logout,
    authError,
    authLoading,

    showAlert,
    hideAlert,
    setNotifications,
    setFriends,
    setPendingInvitations,
    setOnlineUsers,
    setUserNotifcations,
    setGroupChatList,
    resetFriends,

    setChatType,
    setChosenChatDetails,
    setMessages,
    setGroupNotifications,
    addNewMessage,
    resetChat,
    setChosenGroupChatDetails,

    setTyping,
    setInitialTypingStatus,

    setLocalStream,
    setRemoteStream,
    setOtherUserId,
    setAudioOnly,
    setScreenSharingStream,
    setScreenSharing,
    setCallRequest,
    setCallStatus,
    resetVideoChatState,

    openRoom,
    setRoomDetails,
    setActiveRooms,
    setLocalStreamRoom,
    setRemoteStreams,
    setAudioOnlyRoom,
    setScreenSharingStreamRoom,
    setIsUserJoinedWithAudioOnly,
}
// ...
interface setNotifications {
    type: actionTypes.setNotifications;
    payload: Array<OnlineNotifaction>;
}
// ...
interface setUserNotifcations {
    type: actionTypes.setUserNotifcations;
    payload: Array<UserGetNotification>;
}
// ...
export interface SetGroupNotifications {
    type: actionTypes.setGroupNotifications;
    payload: Array<{
        notifications: number;
        groupChatId: string;
    }>;
}

export interface ResetChat {
    type: actionTypes.resetChat;
}

export interface ResetFriends {
    type: actionTypes.resetFriends;
}

export interface Message {
    _id: string;
    content: string;
    author: {
        username: string;
        _id: string;
    };
    createdAt: string;
}
// ...
interface setOtherUserId {
    type: actionTypes.setOtherUserId;
    payload: {
        otherUserId: string | null;
    };
}

interface setScreenSharingStream {
    type: actionTypes.setScreenSharingStream;
    payload: {
        stream: MediaStream | null;
        isScreenSharing: boolean;
    };
}

interface SetAudioOnly {
    type: actionTypes.setAudioOnly;
    payload: {
        audioOnly: boolean;
    };
}

export type ActiveRoom = {
    roomCreator: {
        userId: string;
        username: string;
        socketId: string;
    };
    participants: {
        userId: string;
        socketId: string;
        username: string;
    }[];
    roomId: string;
};